+++ services/workflows-service/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @ballerine/workflows-service
 
+## 0.7.80
+
+### Patch Changes
+
+- Updated dependencies
+  - @ballerine/workflow-core@0.6.76
+  - @ballerine/workflow-node-sdk@0.6.76
+
 ## 0.7.79
 
 ### Patch Changes